pkg/utils: merge lightning/bytesBuffer to utils
